changed the C and E calculations
made a few edits for C and E.  changed ndays to 20 and dt to .01.  Still
havening trouble with the resources persisting past the end of the
season.  I suppose this is possible in real life, but i think it says
something weird about the model.  So, I think there is still something a
little wrong about the within season dynamics.  I may need to revert to
a previous version to see if it was running better before I mucked
around with it....  .