changed the C and E calculations
made a few edits for C and E.  changed ndays to 20 and dt to .01.  Still
having trouble with the resources persisting past the end of the season
(as defined by tsteps).  I suppose this is pobbile in real life, and we
should think about what a realistic drawdown time course looks like, but
for now 5000 tsteps isn't enough for the season to end.  It might say
something weird about the within season dynamics, but I'm not sure yet.
I may need to revert to a previous version to see if it was running
better before I mucked aroudn with it.