big changes:  added ODE solver, pulled out sections and put in source files, restructured within year arrays
There are a lot of changes here.  The first two are the big ones:
1) I pulled out the paramterization sections into separate source files.
These source files were added in a separate commit (6 source files:
getSpecies, getEnvt, getGraphParms, getRunParms, as well as two ode
solver files: NoCompN and ResCompN)
2) I added the ode solver to the loop.  This caused some other changes
because the output from the ode solver is a dataframe with elements
time,R, B1,B2....  I saved teh dataframe for each year as an element in
a list.  Ultimately, this may use too much memory, but it was useful for
now.
3.  I commented out the larger loop whie I was working on the annual
loop
4. I fixed the first couple of figures generating bits to work with the
new within-year data structure,but  didn't finish that fix.  I have
marked where I left off, and it shoudln't be too hard to figure out how
to fix.
5. I changed how we handled the extra time for a nonstationary run.  If
it is a nonstationary run, then nonsta is assigned a value >0 that is
the number of years of nonsttionarity.  This is then used as a flag to
adjust the parameter values in other places.  Thereofre, we cna just
switch the flag (from 0 to number of nonstationary years) to to
stationary or nonstationary runs.  Hope that will work.

Sorry for all the different things that got changed here... poor coding
style/