changes to tauP definition & nonstationary runs
For nonstationarity, made variable nonsta a 2 element vector.  nonsta[1]
is number of years of transition; nonsta[2] is the number of years
simulated at the final level.  Made changes in getEnvt and
PhenologyModel to agree with this.

Also, made it so shifting tauP goes from beta(10,10) to beta(5,15);
i.e., both p and q parameters change over the period nonsta[1].  This
makes the changes in tauP mostly a change in mean from .5 to .25 without
too much skew at the end