changes for batch file processing
This file is a new version of PhenologyModel.R that is set up for
processing batch files.  All run parameters have been moved into
sourcefiles/getRunParms and all graphing and extraneous stuff has been
moved out of this file.  It now writes an .Rdata file out for every run
into /R/output.  To run it on your own desktop, the only changes to be
made are in getRunParms.