text file that gives list of run parameters for batch processing
getInputParms.txt is called by getRunParms when the file is a batch
process.  Setting an environment variable PHEN_RUNNUM in bash allows you
to select the line to run.  This allows the simulations to be (nearly)
fully automated.  Note that some concatenation still has to occur
because one file is written out for each of the tasks the batch array