updated sourcing, relative paths, and order of events
In addition to updating the sourcing and relative paths, the meaningful
change here was to re-order when N and B are updated.  This allows the
data to be output each year more sensibly.  Now, years count from 1:nyrs
and all of the variables are calculated in year y from year y-1 values.
In addition, this means that the order of events is:  (1) rcrt(y) is the
number of seeds being added in y that were genereated in y-1, (2) count
the seeds in year y as the prior year (less germ adn overwinter)
+rcrt(y), (3) within year dynamics generate Bout[y] and Bfin[y].