updated summary so outputs at end of each period
I updated the summary output file so that it write out a new line at the
end of each period.  NOTE that the list of headings has changed a bit to
include PERIOD and NPERIOD between runnum and yout, gi_mean_pre is now
gi_mean, and tauIPns_pre/post is no tauIP_mean.  For a run with only a
stationary period (or if the species go extinct during the first
period), it will write one line.  For runs that have a ST-NS or ST-NS-ST
or ST-ST periods, it will write out one line per period until the
species go extinct.  In a period where extinction occurs midway through,
the g_mean and tauIP_mean are the within-period average up to the point
of extinction.